requirements.txt
# Twitter Bot System Dependencies

# Core dependencies
python-telegram-bot==20.7
twikit==1.3.8
cryptography==41.0.7

# HTTP and networking
aiohttp==3.9.1
requests==2.31.0

# Data handling
pydantic==2.5.2
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional: For better performance
uvloop==0.19.0; sys_platform != "win32"

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
